+++ tests/Request/RegisterRequestBuilderTest.php
@@ -76,11 +76,13 @@ class RegisterRequestBuilderTest extends TestCase
         yield ['whereStartsWith', ['foo', 'baz'], '"foo" LIKE \'baz%\''];
         yield ['whereEndsWith', ['foo', 'baz'], '"foo" LIKE \'%baz\''];
         yield ['wherePartial', ['foo', 'baz'], '"foo" LIKE \'%baz%\''];
+        yield ['whereContains', ['foo', 'baz'], '"foo" LIKE \'%baz%\''];
         yield ['whereLt', ['foo', 'baz'], '"foo" < \'baz\''];
         yield ['whereLte', ['foo', 'baz'], '"foo" <= \'baz\''];
         yield ['whereGt', ['foo', 'baz'], '"foo" > \'baz\''];
         yield ['whereGte', ['foo', 'baz'], '"foo" >= \'baz\''];
         yield ['whereEq', ['foo', 'baz'], '"foo" = \'baz\''];
+        yield ['whereNeq', ['foo', 'baz'], '"foo" != \'baz\''];
         yield ['whereLike', ['foo', 'baz'], '"foo" LIKE \'baz\''];
         yield ['where', ['foo', 'baz', '='], '"foo" = \'baz\''];
     }